+++ src/formatter/html.rs
@@ -1,7 +1,7 @@
 use super::Formatter;
 
 pub trait HtmlFormatter: Formatter {
-    fn pre_tag(&self) -> String;
-    fn code_tag(&self) -> String;
+    fn open_pre_tag(&self) -> String;
+    fn open_code_tag(&self) -> String;
     fn closing_tags(&self) -> String;
 }